switch: Added tcpdump tracing support.
Packets in and out of the switch are globally traced to a pcap output
file. The file is enriched a bit so that it can be used for testing
the operation of the switch.

Extra metadata -- port ID and input vs. output -- is sneakily added
onto each PCAP record after the end of the Ethernet frame. tcpdump and
wireshark seem to be okay with this. (Later this metadata should
become a new libpcap Data Link Type with nice formatting in tcpdump --
but that is then and this is now.)