Added Linux TAP (tuntap) network medium.
TAP is switch ports that connect to the host OS that Snabb Switch is
running on. They make it possible for the host to participate in the
virtual network e.g. to act as a router or connect to VMs.

See http://www.kernel.org/doc/Documentation/networking/tuntap.txt for
more details.