report.lua: Added reporting capability.
It works like this:

If you set report.commanddir = "/tmp/foo/" then the switch will poll
this directory for Lua scripts. When it finds a script S it will load
and execute the script inside the switch, write the output to S.out,
and remove S when finished.

This is a simple form of RPC / remove-eval into the switch.