intel.lua: Created optimized selftest2 that transmits at 35ns/packet.
The trick for better performance is to avoid updating hardware
registers on every packet, specifically the TDT transmit ring tail
pointer register. Now there is an explicit flush_tx() to do this and
meanwhile transmits update a shadow/cache variable called 'tdt'.

This commit is a bit messy. Cleanup to follow.