Fixed crash in selftest when multiple PCI devices are available.
The selftest executes for each suitable PCI device found. As part of
the test, an instance of the intel NIC is instantiated, which caused
the FFI cdefs to be executed more than once resulting in a crash due
to the attempt to redefine the cdefs.

See
https://github.com/SnabbCo/snabbswitch/issues/34#issuecomment-12409461
for an example of the stack trace.