pci.lua, intel10g.lua: Load a dedicated Lua module for each NIC port.
This is an experiment in Lua coding style. Now device driver modules
can be written as if there is only one device to control. The PCI
address of this device is known when the module is loaded.

This is my current favorite of the three main programming styles that
I have seen for using the same driver module for multiple devices:

1. Table-based OO. The discipline of Using `self:foo()` and `self.foo`
   to access everything that is unique for each device.
2. Closure-based OO. Wrapping the whole driver in a `new()` function
   and defining `local` variables for everything that is unique for each device.
3. Module-based OO. The module code is specific to one device and no
   special programming discipline is needed. The module needs to be
   loaded in a special way and reloaded for each instance. (This is
   what this commit implements.)

Feedback welcome!