Tweaked the font choices a bit for consistency
I couldn't live with a sans serif font for the body text, so I've
switched to droid serif for now.  In addition, to maintain consistency
with the family, I've switched the monospaced font used for code
snippets to droid sans mono.

Droid was designed for on-screen use, so these may not be the best for
print, but given that I read a ton of PDFs online, I figured I'd stick
with this one for now.