Made some improvements to the structure for "make book".
The table of contents is now more topic-oriented:

    Startup
    Memory
    PCI
    Hardware Ethernet device drivers
    Linux Vhost_net Virtio-based software Ethernet I/O
    Networking
    Library

The idea is that you can ask yourself which aspect of Snabb Switch you
are interested in, read that one corresponding chapter, and know
everything there is to know about that particular aspect.

This is particularly due to refactoring the general snabb.c "Linux
support" into task-specific files: memory.c, pci.c, etc.