Commented out the enumitem package due to a bug
The LaTeX package `enumitem` contains a bug that prevents us from
producing the final PDF output.  I only used this to lessen the
space in between bulleted and enumerated lists so it's not really
that big of a deal.  I think there may be other ways to accomplish
the same that I can pursue later.