Need to terminate end of a list
Due to the limitation with the manner in which we are programmatically
generating our markdown files from source, we need to terminate any
lists that immediately precede a block of Lua code due to the four
space rule of markdown.

The four space rule states that a list item may contain multiple
blocks of content so long as the next block starts with four spaces.
For example:

```
1. Item 1

    Still item 1 (new paragraph though)

2. Item 2

    Still item 2 (new paragraph though)

    function test(x)
      print(x)
    end
```

In the above example, the function is included as another block
level element of item 2.  It is not typeset as a code block due
to the four space rule.  No matter how many blank lines you
insert between the two, it will still be counted as part of the
prior item.

This means that we _must_ use fenced code blocks with the triple
backticks rather than rely on indentation. I don't have time at
the moment, but we'll have to tweak the method in which we
generate the markdown or we have to let it be known that lists
cannot immediately precede code blocks.