Add app for raw socket access to regular network devices
The RawSocket app provided in src/apps/socket can be used to read and
write bare packets from regular network devices.  It can be used to
access NICs for which no native Snabb Switch driver exists.  It can
also facilitate debugging with, e.g., tcpdump, which is not straight
forward to use with the native driver.  The app is not recommended for
handling traffic with high demands on performance.