Extract the virtio-net device code into a separate module
No features are supported in this implementation
This will result in having the following descriptor layout for both RX/TX:

 - virtio_header
 - data buffer

The buffer origin struct is extended to include the information needed
to access the virtio header.

no-IOMMU - translate all virtual addesses to physical

NOTE: This is slow! Using pread for each packet ...

Signed-off-by: Nikolay Nikolaev <n.nikolaev@virtualopensystems.com>