Add capability to recycle objects to class.lua
The creation of instance objects should be avoided in packet
processing loops to reduce garbage collection overhead.  class.lua now
supports a simple mechanism where an app can declare an object to no
longer be in use.  This object is then put on a list in the
corresponding class and will be re-used by the constructor method
instead of allocating a new object.

The class API has been changed slightly to no longer require the
declaration of constructor methods.  Any method can act as a
constructor or extend the standard constructor by calling the new()
method of the super class.