[apps] add fuzz app
The generation is done in 2 phases:
 - payload generation - stacking Raw data, Ethernet, IPv4, IPv6,
   TCP and UDP protocol layers
 - scatter the data to predefined iovec layouts

Each packet is marked at its payload end with a unique number. Upon
receiving this number is used to find the originating match. Their
parsed headers are then compared and verified.

NOTE: This implies that the the packets are received assembled if
any fragmentation was used. So GSO should be enabled on both TX/RX.

Signed-off-by: Nikolay Nikolaev <n.nikolaev@virtualopensystems.com>