[packet] add scatter()
Add a packet function that does the opposite of the coalescse.
It take a packet and returnes a 'scattered' clone of it.
The second argument is a scatter pattern list. It has the form
{ {len, offset}, {len, offset} ... }

An example invocation:

scattered = packet.scatter(p, {{10}, {100, 20}}

This will create a new packet with the following iovec layout:
 - 10 bytes at offset 0
 - 100 bytes starting at offset 20
 - the rest starting at offset 0

Signed-off-by: Nikolay Nikolaev <n.nikolaev@virtualopensystems.com>