Add simplistic but efficient packet classifer
The module lib/protocol/matcher.lua provides a simple framework to
check whether a chunk of memory contains specific byte-strings of a
certain size at particular offsets.

This is considered to be a hack around the matching mechanism provided
by the datagram:parse() method, which performs poorly with regard to
garbage collection.

A better approach could be to integrate the matching engine of libpcap
into Snabb Switch.