Change of semantics of the datagram:push() method
To allow for more flexibility in the parsing of protocol headers in a
datagram, the parse() method now parses a single header only.  The new
method parse_seq() provides the same semantics as the old parse().

A new method parse_n() has been added to parse exactly n protocol headers.

Also new is the pop_raw() method that pops a protocol header without
parsing it.  This requires the application to already know the type
and size of the next header.