Add functionality to the datagram module
The new method parse_match() allows the parsing and matching of a
single protocol header without having to allocate a table, which can
help to avoid garbage.

The new method parse_n() parses exactly n protocol headers without
applying any additional checks.

Also new is the pop_raw() method that pops a protocol header without
parsing it.  This requires the application to already know the type
and size of the next header.