[virtio] suppress event usage
Notify upper layers only if VRING_F_NO_INTERRUPT is not set in the avail ring's flags.

Setting VRING_F_NO_NOTIFY in the used ring's flags (both tx/rx), should suppress the driver
from genereting interrupts. We polling anyway, so we don't need them.

This is a simple interrupt suppression mechanism, but still we have to implements it.
The more advanced options is to implement EVENT_IDX.

Signed-off-by: Nikolay Nikolaev <n.nikolaev@virtualopensystems.com>