lib.malloc(): Use memory.dma_alloc() instead of C.malloc().
This causes allocations to be made on HugeTLB memory.

This also introduces a limitation: can't allocate an object larger
than one TLB. This is not a problem with 1GB TLB but is a problem with
2MB TLB e.g. when allocating all packets/buffers as an array. If we
keep this change then buffer.lua/packet.lua will need to be updated
to (for example) allocate structs one by one instead of in one big
array.