makefile: remove hardcoded deps/luajit paths
In-tree luajit installation is moved to deps/.install
The dynasm *lua and *h files are copied there too.

All references to ../deps in src/Makefile are replaced with env variables.
Those are exported in the root Makefile.

To link to the dynamic luajit library compile with:
 LUAJIT_DYNAMIC_LINK=1 make

To compile with externally install luajit library:
 LUAJIT_PREFIX=/path/to/luajit/installation make

Signed-off-by: Nikolay Nikolaev <n.nikolaev@virtualopensystems.com>