New semantics for datagram class
The push() method now appends headers to the beginning of the packet
buffer of iovec #0.  This simplifies the contruction of packets and
avoids unnecessary buffer allocations if different apps push to the
same packet.  However, this change is incompatible to the previous
version.

All parse methods are now also available for datagrams created from
scratch to make the API more consistent and to avoid checks for
special cases.

A new constructor reuse() allows re-using an instance without putting
it on the freelist first.