Make nfv application detect NIC driver
Previously, the nfv application assumed that the NIC was an Intl NIC.
This patch changes it so that uses the string returned by
pci.which_driver() to require the driver module for the NIC.  The
driver module is now assumed to have a "driver" property that returns
the driver class which is then instantiated with new().  Both the
Intel and the Solarflare drivers are adapted for the new scheme.