engine.configure(): Fix order of operations during config update
Ensure that 'stop' actions are executed before 'start' actions.

Otherwise there can be problems when a new app and an old app both
want to use the same resource and the new app starts before the old
one has terminated. For example, when replacing an app like intel10g
that needs exclusive use of a PCI device.