Have nd_light overwrite ethernet headers instead of adding them
In order to maintain symmetry, the semantics of nd_light are changed
such that it expects full Ethernet frames on its north port.  Instead
of adding the header, it merely overwrites it.

The rationale for this change is as follows.  It is desrieable to keep
the Ethernet header in the packets received from the south port when
transmitting them to the north port, which will allow the receiving
app to inspect the entire packet.  In particular, this is necessary if
BPF filters need to be applied to the packet higher up in the app
chain.  To obey the principle of least astonishment and to keep the
interface symmetric, the northbound apps must make sure that the
packets passed to nd_light contain an Ethernet header as well, but the
contents is irrelevant.