intel10g: Keep RXLNGTHERREN disabled
That feature will automatically drop packets that seem to be the wrong
length. This complicates testing when we would quite like to be able
to send garbage packets. There is no obvious benefit for us.

basic_apps.Source also simplified to skip setting an IP protocol. That
was only a workaround to prevent his feature from dropping packets.