Improve efficiency of lib/ipc/shmem mechanism
The new dictionary() method of the base class allows for more
efficient access to objects in a shared memory segment by
pre-computing pointers after all object have been registered.

The commit also includes a fix for the counter module and an
implementation of the BITS construct in the mib module.