Add IPC method using file-backed shared memory
lib/ipc/shmem/shmem.lua implements a class that allows the sharing of
a collection of cdata objects with other processes through a region of
shared memory.

This commit includes two sub classes. The mib subclass implements data
types based on the SMIv2 specification.  It is intended to provide an
interface to a SNMP sub-agent running on the same host.

The counter subclass provides a collection of unsigned 64-bit numbers,
intended to be used as generic counters (e.g. byte or packet counters
for interfaces or inter-app links).