Make ./snabb binary more portable
This will enable us to release Snabb Switch as a single executable
file for any modern Linux/x86_64.

./snabb will link with any GLIBC version >= 2.7 (released 2007).  This
makes the binary linker-compatible with older Linux distributions such
as Ubuntu 10.04 and CentOS 6.

Tested when built on Ubuntu 14.04. Building on other distributions may
yield different results, but a selftest script is included to catch that.

(There may be other compatibility issues with older kernels and
CPUs. This only takes care of the dynamic linking.)