Rudimentary support of the ifTable MIB for intel10g
An instance of the intel10g driver creates a row in the ifTable of the
IF-MIB containing just the objects ifDescr, ifType, ifSpeed,
ifAdminStatus, ifOperStatus and ifLastChange to allow monitoring of
the most basic properties and link up/down events.

The mib ipc module requires support in the SNMP agent that serves the
IF-MIB on the local host.