Add ljsyscall to Snabb Switch
This will enable us to rewrite most of our C glue code in Lua.

ljsyscall is an extensive LuaJIT FFI interface towards the kernel:
  https://github.com/justincormack/ljsyscall

The source code is added as a submodule in deps/ljsyscall.

The top-level Makefile copies required source files directly into
src/. Only the code for Linux/x86-64 is imported.

The top-level .gitignore file matches on the copied files.