solarflare: Sink FFI pointer allocations in pull()
Limit the lifetime of local variables assigned to FFI pointers. This
allows allocation of the boxed pointers to be "sunk" and thus prevents
heap allocation. (In particular, prevent the scope of pointer-valued
variables from spanning across conditions and loops that might cause
JIT trace exits where pointers will spill out of registers and get
boxed.)

The pull and push methods seem to be free of allocation now.