src/Makefile: make ./snabb default target and add '.inc' files
'make' will now build only ./snabb. If you want to create files in
bin/ (e.g. bin/snabbnfv) then use 'make bin/<program>' or 'make all'.

Files named *.inc are now (inc)luded in the executable.

Specifically, the contents of a file named

    src/foo/bar.inc

can be loaded as a string with

    require("foo.bar_inc")

This is useful in several ways:

- Command-line documentation can be written into a README file
  (visible on Github) and this can be included in the binary.

- Shell scripts can be written in .sh files and embdedded.
  (This is handy for snabbnfv shell scripts at the moment.)

- Version information could be generated and included too.