memory.c: Fill new allocated DMA memory with zeros
This is to reduce the chance of hard-to-reproduce errors that depend
on initial memory contents.

This happened once upon a time when the NIC was doing more packet
inspection than expected and rejecting some packets that contained
uninitialized garbage payload. (See commit bc04a11.) The patch is
simply intended to reduce the risk of confusing problems.