New usage: packetblaster replay <file> <port...>
packetblaster (previously 'loadgen') now takes a new first argument
saying how to operate. The only currently valid value is 'replay' to
load packets from a pcap file (the traditional behavior).

This is to make it easier to add new behaviors with different keywords
in the future. For example, to generate synthetic traffic, generate
tcp/udp sessions, and so on.