Remove garbage from core.apps.restart_dead_apps()
The "action" table is allocated each time the function is called,
which creates a fair amount of garbage, even if the restart mechanism
isn't actually used, which is the default.

This commit includes a bypass of the function if use_restart is false.
If use_restart is true, the action table is allocated statically and
only the sub-tables that are used by the function are cleared.

The dead-app detection code probably contains more of this kind of
garbage, e.g. in the apply_config_action() function.