Bypass core.apps.restart_dead_apps() if restart mechanism is disabled
The "action" table is allocated each time the function is called,
which creates a fair amount of garbage, even if the restart mechanism
isn't actually used, which is the default.

This commit includes a bypass of the function if use_restart is false.

The dead-app detection code probably contains more of this kind of
garbage, e.g. in the apply_config_action() function.