lib.checksum: Updated checksum library
The new interface is:

    checksum.ipsum(pointer, length, initial) => checksum

and the most appropriate checksum routine for the running CPU will be
used automatically. There are currently three checksum routines: plain
x86 (imported from DPDK), SSE2 (Tony Rogvall), and AVX2 (my port of
the SSE2 one).

This is not benchmarked very extensively but the first peek at
cycles-per-packet (where packet is really just an array being
benchmarked) is like this:

    PKTSIZE    REF  SSE  AVX2
	 64     26   37    39
	100     39   49    47
	256     83   65    53
	300    128   82    67
	512    182  123    76
	600    212  130   100
       1000    347  210   135
       1500    529  296   189

... and in light of this I made the reference non-SIMD one be used
when size is < 128.