net_device: Added temporary fake checksum offload on receive
The checksum offload is "fake" because it assumes the checksum was
correct without actually testing the result. This has to be fixed by
making it smart enough to look at the IP/UDP/TCP header.

It is still useful for performance measurements because it does make
the CPU checksum the packet data.

This has to be fixed with priority because otherwise a packet with a
genuinely incorrect checksum will not be dropped by the guest. (This
fake implementation is not suitable to be included in a release.)