core.memory: Replace mlockall() with on-demand mlock()
Snabb Switch previously called mlockall() to force all memory in the
process address space to be locked to its physical location. This made
it possible to use all memory for DMA, including memory mapped from
virtual machines.

This had several side-effects:

* Prevent Snabb Switch from being swapped out. (Nice?)
* Prevent VMs that we serve from being swapped out. (Overkill?)
* Requires root permissions at startup. (Overkill?)

The new behavior is to individually mlock() the HugeTLB pages that we
allocate for DMA. The rest of the memory - Snabb Switch and VMs - is
left unlocked. This is feasible now that we only do DMA within this
memory. This also allows us to defer the root-privileged system call
until we need DMA memory, which may not happen for ceratin
applications.