core: Count "frees per second" (fps) and do dynamic idling
"Frees per second" (fps), literally the number of calls to
packet.free() per second, is introduced as a new basic metric for
measuring work done by a Snabb Switch process.

engine.Hz = false now uses a dynamic mechanism to decide how long to
sleep for depending on how busy/idle the process is, as observed from
fps. This is now the default setting and should be suitable for normal
usage.

Load on the process can be printed by calling
engine.report_load(). The load report looks like this:

    load: time: 1.00s  fps: 2,025,221   fpb: 304  bpp: 1280 sleep: 0   us

and the metrics mean:

    time - period of time that the metrics were collected over
    fps  - frees per second
    fpb  - frees per breath
    bpp  - bytes per packet (average packet size)