engine: Add new metric: fpGbps (freed physical Gbps)
fpGbps estimates the physical 10GbE capacity required to carry the
packets that are being freed per second. This is counted from the size
in bytes of the packets being freed and accounting for 10GbE
overhead (minimum packet size, CRC, inter-packet gap).

The report format is now:

    load: time: 1.00s  fps: 2,114,697 fpGbps: 22.570 fpb: 293 bpp: 1325 sleep: 0   us

This is mostly useful for estimating link utilization. If you are
testing with two 10GbE links and your fpGbps is 20 then you are at the
theoretical maximum.

Practically speaking fpGbps is for estimating how close the process is
performing to "line rate" for its workload i.e. the maximum traffic
that can be carried over a physical link. (So you don't have to
memorize that if you are testing with 256-byte packets your line rate
is approx 4.45 Mpps, etc.)

Note: Measuring this when packets are freed my not be ideal because
encapsulations added and removed may be miscounted. In this case the
fpGbps metric would be superceded by a more practical alternative in
the future.