Support configurable MTU in Intel82599 driver
The MTU is now configurable individually for each instance of the
driver.  The MTU includes the L2 header, but not the CRC.  The driver
enables jumbo frames if the MTU exceeds 1514 bytes.  Receive buffers
must be large enough to hold a full-sized packet.  Packets which
exceed the MTU are discarded by transmit().