replace the packet_filter code generator
replaces the 'repeat ... until false' pseudo loops with simple
'if cond1 and cond2 .... then ... return true end'.  hopefully it will be
easier on the JIT; all tests so far it all stay 100% compiled.

there are fewer opportunities to avoid unneeded work, but compensates
by not repeating packet field peeking; at the start we fill out an ffi
spec structure (the same used for stateful tracking) and use that for checking.