vhost_user: Add persistent cache of negotiated features
Make reconnecting to an existing vhost-user socket more robust by
caching the negotiated features (in a /tmp file) instead of relying on
QEMU to remember them. This makes it possible to restart the Snabb
Switch continuously and always successfully reattach to the VMs.

QEMU does not report features reliably and the root cause is that
reconnecting the socket (restarting the switch process) is currently
considered unsupported in QEMU upstream (though we have a small patch
to make it possible).

Specifically:

* QEMU 2.1 often returns the correctly negotiated features on
  reconnect, but occasionally reports a garbage value instead.

* QEMU master returns zero for the negotiated features on reconnect.