PcapFilter app replacing PacketFilter.
This app is simple: it filters packets using one pflua filter
rule. Optionally it also uses conntrack for stateful tracking.

The unit tests are weak: they are intended to detect breakages and
unexpected changes in the behavior of pflua and conntrack, but not to
actually test the correctness of those libraries (which are other units).