Add attach() method to the shmem IPC mechanism
The new method allows a Snabb process to attach to an existing memory
segment as a "slave".  Only a single process is allowed to be the
"master" of a segment.  The master creates a segment via the new()
constructor, while a slave attaches to it with the attach()
constructor.  Both, master and slave have to register() the proper
data types for every object in the segment (unless the segment
contains only objects of a single, well-known type).