apps/packet_filter: Rename pcap_filter.PcapFilter to bpf.BPF
Give a new shorter name to the pflua based packet filter app.

The name BPF is a reference to the Berkeley Packet Filter where the
pflang filter language originated (before being adopted by
libpcap).

  http://en.wikipedia.org/wiki/Berkeley_Packet_Filter
  http://www.tcpdump.org/papers/bpf-usenix93.pdf

This name is somewhat ambiguous in that BPF can refer to a larger body
of software including the /dev/bpf packet capture device
on *BSD. However, the name Pcap has the same ambiguity of pcap-filter
vs libpcap, and networking people seem to prefer "BPF" to "PcapFilter".