Rewrote top-level README.md
The original README is a kind of programming rant/manifesto. That was
pretty suitable at the time it was written, but now we need to explain
what our project is about and in enough detail that people can related
it to other projects such as OVS, DPDK, NETMAP, PF_RING, etc.

This new README attempts to briefly tell people who we are, what we
are doing, what we have done, and how to join.