[draft] Create new 'snabb' module to define the core API
The idea is to explicitly define our core API in one place.  Then
users will not required a lot of individual modules (core.packet,
core.link, etc) but instead access everything through a unified and
well-defined module (snabb).

This draft API would make code look more like this:

    local config = snabb.config()
    config:add_app(...)
    config:add_link(...)
    snabb.engine.configure(config)

and

    local packet = snabb.packet()
    packet:shiftleft(42)
    print("packet length after shortening: " .. packet:length())

Notably you would not have to require() any modules anymore. (You
could require("snabb") if you want to but we could probably just make
it global.)

This MIGHT also make it easier to create new versions of the API with
semantic versioning (http://semver.org/) and even to allow apps to say
which API version they depend on so that we could keep compatibility
with older versions for as long as we want.