spray: Getting Started spray program and sprayer app
This is an experiment in programming style. The programs are written
realistically but have not been tested. The idea is to see if this is
the way we want people's first programs to look.

Compared with the master branch the intended differences are:

- Use the new Snabb API from this branch.
- Use require("snabb1") to explicitly depend on API version 1.
- Use the "apps have a private sandbox" model in sprayer.lua.
- Use the 'link:transmit(packet)' object-oriented syntax variant.
- Make program/ module files simply run instead of loading first.
- Added engine:run({duration = 'toidle'}) option (run until idle).

The overall effect should be shorter code that looks more like plain
Lua scripts.

The use of 'link:transmit(packet)' syntax feels right to me in this
example. It is also a retreat from the position that I took on
snabb-devel where I advocated for 'l:link_transmit(packet)' (after
having previously advocated for link.transmit(l, p)).