Sandbox level 2: one LuaState VM for each app
lib.lua.statebox implements the LuaState handling code using FFI calls
to the Lua C API.  Handles passing simple values and non-recursive tables.
Any userdata or FFI cdata value is cast into a lightuserdata.

core.app_loader uses lib.lua.statebox to create app-compatible objects.
core.config calls this loader when the app is specified with a dotted.name
instead of the usual class table.

Translated apps.basic.source/sink to the new style.  It's very similar to
the sandbox level 1 style, but now it's safe to call require() to reuse
code within the app.  Usually there's no need to call module(), since
we do want most of the code and variables to be 'global' within the
sandbox.  Global functions and some variable names (currently zone and name)
are copied to the app proxy object.

The start code within the sandbox sets the LuaJIT profiler, probably the
app:stop() hook is not the right place to close it.